+++ apps/explorer/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @walkeros/explorer
 
+## 3.0.2
+
+### Patch Changes
+
+- @walkeros/core@3.0.2
+- @walkeros/collector@3.0.2
+- @walkeros/web-source-browser@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @walkeros/core@3.0.1
+- @walkeros/collector@3.0.1
+- @walkeros/web-source-browser@3.0.1
+
 ## 3.0.0
 
 ### Minor Changes