+++ apps/walkerjs/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @walkeros/walker.js
 
+## 3.0.2
+
+### Patch Changes
+
+- @walkeros/core@3.0.2
+- @walkeros/collector@3.0.2
+- @walkeros/web-core@3.0.2
+- @walkeros/web-source-browser@3.0.2
+- @walkeros/web-source-session@3.0.2
+- @walkeros/web-source-datalayer@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @walkeros/core@3.0.1
+- @walkeros/collector@3.0.1
+- @walkeros/web-core@3.0.1
+- @walkeros/web-source-browser@3.0.1
+- @walkeros/web-source-session@3.0.1
+- @walkeros/web-source-datalayer@3.0.1
+
 ## 3.0.0
 
 ### Patch Changes