+++ packages/core/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @walkeros/core
 
+## 3.0.2
+
+## 3.0.1
+
 ## 3.0.0
 
 ### Major Changes