+++ packages/server/transformers/file/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @walkeros/server-transformer-file
 
+## 3.0.2
+
+### Patch Changes
+
+- @walkeros/core@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @walkeros/core@3.0.1
+
 ## 3.0.0
 
 ### Minor Changes