+++ packages/stores/memory/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @walkeros/store-memory
 
+## 3.0.2
+
+### Patch Changes
+
+- @walkeros/core@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @walkeros/core@3.0.1
+
 ## 3.0.0
 
 ### Minor Changes