+++ packages/web/core/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @walkeros/web-core
 
+## 3.0.2
+
+### Patch Changes
+
+- @walkeros/core@3.0.2
+
+## 3.0.1
+
+### Patch Changes
+
+- @walkeros/core@3.0.1
+
 ## 3.0.0
 
 ### Patch Changes